Made Endless Mode available

+++ src/com/narayan/a2048/MainActivity.java
@@ -77,21 +77,21 @@ public class MainActivity extends Activity {
         for (int xx = 0; xx < field.length; xx++) {
             for (int yy = 0; yy < field[0].length; yy++) {
                 if (field[xx][yy] != null) {
-                    editor.putInt(xx + " " + yy, field[xx][yy].getValue());
+                    editor.putInt(xx + " " + yy, (int)field[xx][yy].getValue());
                 } else {
                     editor.putInt(xx + " " + yy, 0);
                 }
 
                 if (undoField[xx][yy] != null) {
-                    editor.putInt(UNDO_GRID + xx + " " + yy, undoField[xx][yy].getValue());
+                    editor.putInt(UNDO_GRID + xx + " " + yy, (int)undoField[xx][yy].getValue());
                 } else {
                     editor.putInt(UNDO_GRID + xx + " " + yy, 0);
                 }
             }
         }
-        editor.putLong(SCORE, view.game.score);
-        editor.putLong(HIGH_SCORE, view.game.highScore);
-        editor.putLong(UNDO_SCORE, view.game.lastScore);
+        editor.putInt(SCORE, view.game.score);
+        editor.putInt(HIGH_SCORE, view.game.highScore);
+        editor.putInt(UNDO_SCORE, view.game.lastScore);
         editor.putBoolean(CAN_UNDO, view.game.canUndo);
         editor.putInt(GAME_STATE, view.game.gameState);
         editor.putInt(UNDO_GAME_STATE, view.game.lastGameState);
@@ -126,9 +126,9 @@ public class MainActivity extends Activity {
             }
         }
 
-        view.game.score = settings.getLong(SCORE, view.game.score);
-        view.game.highScore = settings.getLong(HIGH_SCORE, view.game.highScore);
-        view.game.lastScore = settings.getLong(UNDO_SCORE, view.game.lastScore);
+        view.game.score = settings.getInt(SCORE, view.game.score);
+        view.game.highScore = settings.getInt(HIGH_SCORE, view.game.highScore);
+        view.game.lastScore = settings.getInt(UNDO_SCORE, view.game.lastScore);
         view.game.canUndo = settings.getBoolean(CAN_UNDO, view.game.canUndo);
         view.game.gameState = settings.getInt(GAME_STATE, view.game.gameState);
         view.game.lastGameState = settings.getInt(UNDO_GAME_STATE, view.game.lastGameState);
